backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database and ORM
prisma==0.11.0
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver
redis==5.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and Testing
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring and Logging
structlog==24.1.0
